requirements.txt
torch>=2.0.0
diffusers>=0.20.0
transformers>=4.30.0
peft>=0.10.0
accelerate>=0.20.0
torchvision>=0.15.0
pillow>=9.0.0
numpy>=1.20.0
tqdm>=4.60.0
scikit-learn>=1.0.0
pyyaml>=6.0
matplotlib>=3.5.0